+++ src/components/BatchItemClone.svelte
